Börjat kolla på extensionmetoder till Booking. Hur koppla ihop Customer och Vehicle i Booking?

+++ Common/Classes/Motorcycle.cs
@@ -13,12 +13,12 @@ public class Motorcycle : IVehicle
     public string RegNo { get; set; } = string.Empty;
     public string Brand { get; set; } = string.Empty;
     public int Odometer { get; set; }
-    public int CostKm { get; set; }
+    public double CostKm { get; set; }
     public VehicleTypes VType { get; set; }
     public int DailyCost { get; set; }
     public VehicleStatus Status { get; set; } = default;
 
-    public Motorcycle(string regNo, string brand, int odometer, int costKm, VehicleTypes vType,
+    public Motorcycle(string regNo, string brand, int odometer, double costKm, VehicleTypes vType,
         int dailyCost, VehicleStatus status)
     {
         RegNo = regNo;
