+++ Common/Interfaces/IVehicle.cs
@@ -7,7 +7,7 @@ namespace Common.Interfaces
         public string RegNo { get; set; }
         public string Brand { get; set; }
         public int Odometer { get; set; }
-        public int CostKm { get; set; }
+        public double CostKm { get; set; }
         public VehicleTypes VType { get; set; }
         public int DailyCost { get; set; }
         public VehicleStatus Status { get; set; }
